Commit of 2/12/2024
God blessed us to get the search bar fully functional. Also changed some js around in the script.js file so that it stops making hundreds of errors. Moved that code to the index-script.js file which is where the code should be as it only is used for the index page.